# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: TTJobs Backend API
  version: "1.0.0"
  description: |
    API contract for auth, users, companies, jobs, and job applications.
    This file is static and can be imported into Swagger Editor or Postman.
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Users
  - name: Companies
  - name: Jobs
  - name: Applications
  - name: Notifications
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "200":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          description: Validation error
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: JWT token string
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Invalid credentials
  /api/users/info:
    get:
      tags: [Users]
      summary: User test endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info test
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/admin/test:
    get:
      tags: [Users]
      summary: Admin test endpoint
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin access test
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/companies:
    get:
      tags: [Companies]
      summary: List companies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyDTO'
    post:
      tags: [Companies]
      summary: Create company (Recruiter/Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyWriteRequest'
      responses:
        "200":
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDTO'
        "403":
          description: Forbidden
  /api/companies/{id}:
    get:
      tags: [Companies]
      summary: Get company by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDTO'
        "404":
          description: Company not found
    put:
      tags: [Companies]
      summary: Update company (Owner recruiter/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyWriteRequest'
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDTO'
        "403":
          description: Forbidden
    delete:
      tags: [Companies]
      summary: Delete company (Owner recruiter/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company deleted
        "403":
          description: Forbidden
  /api/companies/{companyId}/members:
    get:
      tags: [Companies]
      summary: Get company members (owner/admin member/system admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyMemberDTO'
        "403":
          description: Forbidden
    post:
      tags: [Companies]
      summary: Add member to company (owner/admin member/system admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyMemberUpsertRequest'
      responses:
        "200":
          description: Company member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyMemberDTO'
        "400":
          description: Invalid role or invalid user type
        "409":
          description: Duplicate member
  /api/companies/{companyId}/members/{memberId}:
    put:
      tags: [Companies]
      summary: Update company member role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyMemberUpsertRequest'
      responses:
        "200":
          description: Company member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyMemberDTO'
    delete:
      tags: [Companies]
      summary: Remove company member
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company member removed
        "400":
          description: Owner cannot be removed or last admin cannot be removed
  /api/jobs:
    get:
      tags: [Jobs]
      summary: Get candidate-facing jobs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDTO'
    post:
      tags: [Jobs]
      summary: Create job (Recruiter/Admin with company ownership)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobWriteRequest'
      responses:
        "200":
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        "403":
          description: Forbidden
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        "404":
          description: Job not found
    put:
      tags: [Jobs]
      summary: Update job (Owner recruiter/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobWriteRequest'
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
        "403":
          description: Forbidden
    delete:
      tags: [Jobs]
      summary: Archive job (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Archived successfully
  /api/jobs/{id}/close:
    put:
      tags: [Jobs]
      summary: Close job
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDTO'
  /api/jobs/company/{companyId}:
    get:
      tags: [Jobs]
      summary: Get jobs of one company (owner recruiter/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Company jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDTO'
        "403":
          description: Forbidden
  /api/jobs/search:
    get:
      tags: [Jobs]
      summary: Search jobs with filters and pagination
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: title
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: companyName
          schema:
            type: string
        - in: query
          name: jobType
          schema:
            type: string
        - in: query
          name: experienceLevel
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            example: open
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobDTO'
  /api/applications:
    get:
      tags: [Applications]
      summary: Get all applications (Admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobApplicationDTO'
        "403":
          description: Forbidden
  /api/applications/user/{userId}:
    get:
      tags: [Applications]
      summary: Get applications by user (self/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobApplicationDTO'
  /api/applications/job/{jobId}:
    get:
      tags: [Applications]
      summary: Get applications by job (owner recruiter/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Job applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobApplicationDTO'
  /api/applications/recruiter/my-jobs:
    get:
      tags: [Applications]
      summary: Get applications for jobs owned by current recruiter
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recruiter applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobApplicationDTO'
        "403":
          description: Forbidden
  /api/applications/apply:
    post:
      tags: [Applications]
      summary: Apply for a job (Candidate for self only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: jobId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplicationDTO'
        "400":
          description: Invalid apply condition
        "403":
          description: Forbidden
        "409":
          description: Duplicate application
  /api/applications/{applicationId}/status:
    put:
      tags: [Applications]
      summary: Update application status (Recruiter owner/Admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          required: true
          schema:
            type: string
            example: reviewing
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplicationDTO'
        "400":
          description: Invalid transition
        "403":
          description: Forbidden
  /api/applications/{applicationId}/withdraw:
    put:
      tags: [Applications]
      summary: Withdraw application (owner candidate only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Withdraw success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplicationDTO'
        "400":
          description: Invalid terminal transition
        "403":
          description: Forbidden
  /api/applications/{applicationId}:
    delete:
      tags: [Applications]
      summary: Legacy withdraw endpoint (soft withdraw)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Withdraw success
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Get my notifications
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
  /api/notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notifications count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  unreadCount:
                    type: integer
                    format: int64
  /api/notifications/{notificationId}/read:
    put:
      tags: [Notifications]
      summary: Mark one notification as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Marked as read
        "404":
          description: Notification not found
  /api/notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications marked as read

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Thinh
        email:
          type: string
          format: email
          example: test1@gmail.com
        password:
          type: string
          minLength: 6
          example: "123456"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [CANDIDATE, RECRUITER, ADMIN]
        phone:
          type: string
        address:
          type: string
        experienceYears:
          type: integer
        cvUrl:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        website:
          type: string
        industry:
          type: string
    CompanyWriteRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        website:
          type: string
        industry:
          type: string
        logoUrl:
          type: string
    CompanyMemberDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userEmail:
          type: string
        memberRole:
          type: string
          enum: [RECRUITER, ADMIN]
        createdAt:
          type: string
          format: date-time
    CompanyMemberUpsertRequest:
      type: object
      required: [userId, memberRole]
      properties:
        userId:
          type: integer
          format: int64
        memberRole:
          type: string
          enum: [RECRUITER, ADMIN]
    JobDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        location:
          type: string
        salary:
          type: number
          format: double
        salaryMin:
          type: number
          format: double
        salaryMax:
          type: number
          format: double
        currency:
          type: string
          example: VND
        jobType:
          type: string
        experienceLevel:
          type: string
        status:
          type: string
          enum: [draft, open, closed, archived]
        postedDate:
          type: string
          format: date-time
        applicationDeadline:
          type: string
          format: date-time
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
    JobWriteRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        salary:
          type: number
          format: double
        salaryMin:
          type: number
          format: double
        salaryMax:
          type: number
          format: double
        currency:
          type: string
          example: VND
        jobType:
          type: string
        experienceLevel:
          type: string
        status:
          type: string
          enum: [draft, open, closed, archived]
        applicationDeadline:
          type: string
          format: date-time
        company:
          type: object
          properties:
            id:
              type: integer
              format: int64
    JobApplicationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        applicationDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [submitted, reviewing, shortlisted, interviewed, offered, hired, rejected, withdrawn]
        userId:
          type: integer
          format: int64
        userName:
          type: string
        jobId:
          type: integer
          format: int64
        jobTitle:
          type: string
        companyName:
          type: string
    NotificationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time